requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
asyncpg
psycopg2-binary
alembic
pydantic
python-dotenv